+++ CHANGELOG.md
@@ -13,3 +13,5 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Changed
 
+- fix/1201967437941088 - OrchestratorBuilder lacks container dump options
+